+++ ogre/include/ignition/rendering/ogre/OgreDepthCamera.hh
@@ -98,6 +98,9 @@ namespace ignition
           std::function<void(const float *, unsigned int, unsigned int,
           unsigned int, const std::string &)>  _subscriber) override;
 
+      // Documentation inherited.
+      public: virtual void PreRender() override;
+
       /// \brief Implementation of the render call
       public: virtual void Render() override;
 
@@ -142,6 +145,9 @@ namespace ignition
 
       /// \brief Create the camera.
       protected: void CreateCamera();
+
+      /// \brief Create point cloud texture. This stores xyz rgb data
+      private: void CreatePointCloudTexture();
 
       /// \brief Communicates that a frams was rendered
       protected: bool newData = false;